Fix documentation for updating the email provider (#220)
* Fix documentation for updateEmailProvider

Add a 'params' argument as the first parameter of the updateEmailProvider function. Calling management.updateEmailProvider(data, cb) leads to a 'ArgumentError: The data must be an object' and causes arguments mismatch.

* Fix documentation for emailProvider.update

Add a 'params' argument as the first parameter of the updateEmailProvider function, relates to the previous commit. Add missing 'data' parameter in the example of function call.

* Fix documentation for updateRule